[ALPS04016653] DISP: fix build error
[Detail]
fix build error in ARM 32bit project

MTK-Commit-Id: dae96ecf72ce9ff14e79950f9fbde12ec0ba2159

CR-Id: ALPS04016653
Feature: Display Driver (DSI Interface)
(cherry-pick from 07017bf49b196e23fa67265b673f1bb8979df033)

+++ drivers/misc/mediatek/video/mt6771/videox/debug.h
@@ -44,13 +44,13 @@ extern struct fps_debug fps_info_debug;
 
 
 /* show hrt */
-extern int hrt_high, hrt_low;
+extern unsigned long long hrt_high, hrt_low;
 extern int hrt_show_flag;
 
 /* show layer_en num*/
 extern int layer_en_num;
 extern int layer_en_num_flag;
-extern int layer_size_high;
+extern uint64_t layer_size_high;
 extern int layer_size_low;
 extern int layer_size_flag;
